requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.2
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cobble==0.1.4
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.2
djangorestframework==3.16.0
docx2pdf==0.1.8
fire==0.7.0
fonttools==4.57.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.3
lxml==5.4.0
mammoth==1.9.0
numpy==2.2.5
opencv-python-headless==4.11.0.86
pandoc==2.4
pdf2docx==0.5.8
pdfminer.six==20250416
pillow==11.2.1
plumbum==1.9.0
ply==3.11
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyMuPDF==1.25.5
pyOpenSSL==25.0.0
pypandoc==1.15
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
# pywin32==310
redis==5.2.1
service-identity==24.2.0
setuptools==80.0.0
six==1.17.0
sqlparse==0.5.3
termcolor==3.1.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
